Cast within assertion on double

+++ include/albatross/src/utils/vector_utils.hpp
@@ -16,7 +16,8 @@
 namespace albatross {
 
 inline std::size_t safe_cast_to_size_t(double x) {
-  ALBATROSS_ASSERT(x < std::numeric_limits<std::size_t>::max());
+  ALBATROSS_ASSERT(
+      x < static_cast<double>(std::numeric_limits<std::size_t>::max()));
   return static_cast<std::size_t>(x);
 }
